# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نظام المكتبة الجامعية",
        "description": "توثيق API لنظام المكتبة الجامعية",
        "contact": {
            "email": "admin@university.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل الدخول للطالب",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "student_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": " يوسف طارق احمد"
                                    },
                                    "student_id": {
                                        "type": "numeric",
                                        "example": 15901
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تسجيل الدخول بنجاح."
                    },
                    "401": {
                        "description": "بيانات الطالب غير صحيحة."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل خروج الطالب",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "تم تسجيل الخروج بنجاح."
                    },
                    "401": {
                        "description": "غير مصرح لك."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "الحصول على بيانات الطالب الحالي",
                "description": "إرجاع بيانات الطالب المسجل دخوله حالياً",
                "operationId": "8e0149b7f62d4e6ff6573491d62b4560",
                "responses": {
                    "200": {
                        "description": "بيانات الطالب",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "يوسف طارق احمد"
                                        },
                                        "student_id": {
                                            "type": "integer",
                                            "example": 15901
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://localhost/storage/students/image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح لك"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/books/{title}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "عرض تفاصيل كتاب",
                "description": "عرض تفاصيل كتاب معين باستخدام عنوان الكتاب",
                "operationId": "6ba41c29095cb5c729bc18840390e916",
                "parameters": [
                    {
                        "name": "title",
                        "in": "path",
                        "description": "عنوان الكتاب",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تفاصيل الكتاب",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "quantity": {
                                            "type": "integer"
                                        },
                                        "category": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "authors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح"
                    },
                    "404": {
                        "description": "الكتاب غير موجود"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/qr-login": {
            "post": {
                "tags": [
                    "Library Access"
                ],
                "summary": "تسجيل دخول الطالب للمكتبة عن طريق QR",
                "operationId": "afcbef53ad72aa6b9babd112ddbd1830",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_content"
                                ],
                                "properties": {
                                    "qr_content": {
                                        "type": "string",
                                        "example": "library_qr_code"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تسجيل الدخول بنجاح"
                    },
                    "400": {
                        "description": "رمز QR غير صالح"
                    },
                    "401": {
                        "description": "غير مصرح له"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}